backend2.0/requirements.txt
fastapi
uvicorn
pydantic
pyjwt
argon2-cffi
aiosqlite
databases